pyproject.toml
[tool.poetry]
name = "swolbot"
version = "0.1.0"
description = "Pyhton Discord bot to track deaths and reward you with exercise"
authors = ["Dan Rubin <dmrubin3@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
"discord.py" = "^1.5.1"
loguru = "^0.5.3"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pre-commit = "^2.8.2"
pylint = "^2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
